+++ app/src/model/Customer.java
@@ -4,26 +4,27 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Customer extends User {
-	private List<Ticket> managedTickets;
+	private List<Ticket> createdTickets;
 
+	
 	public Customer(){}
 	
 	public Customer(String name, String surname, String email, String password, ProfilePicture profilePhoto) {
 		super(name, surname, email, password, profilePhoto);
-		this.managedTickets = new ArrayList<Ticket>();
+		this.createdTickets = new ArrayList<Ticket>();
 	}
 
-	public List<Ticket> getManagedTickets() {
-		return managedTickets;
+	public List<Ticket> getCreatedTickets() {
+		return createdTickets;
 	}
 
-	public void setManagedTickets(List<Ticket> managedTickets) {
-		this.managedTickets = managedTickets;
+	public void setCreatedTickets(List<Ticket> createdTickets) {
+		this.createdTickets = createdTickets;
 	}
 
 	@Override
 	public String toString() {
-		return "Customer [managedTickets=" + managedTickets + "]"+ "]";
+		return "Customer [createdTickets=" + createdTickets + "]" +"]";
 	}
